+++ src/zope/meta/config_package.py
@@ -22,6 +22,7 @@ import tomlkit
 from packaging.version import InvalidVersion
 from packaging.version import parse as parse_version
 
+from .set_branch_protection_rules import get_package_name
 from .set_branch_protection_rules import set_branch_protection
 from .shared.call import abort
 from .shared.call import call
@@ -156,6 +157,12 @@ def handle_command_line_arguments():
         help='Define a git branch name to be used for the changes. '
         'If not given it is constructed automatically and includes '
         'the configuration type')
+    parser.add_argument(
+        '--no-admin',
+        dest='admin',
+        action='store_false',
+        default=True,
+        help='Do not try to do steps that require GitHub admin rights.')
 
     args = parser.parse_args()
     return args
@@ -750,20 +757,18 @@ class PackageConfiguration:
                 if self.args.push:
                     call('git', 'push', '--set-upstream',
                          'origin', self.branch_name)
-            print()
-            print('If you are an admin and are logged in via `gh auth login`')
-            print('update branch protection rules? (y/N)?', end=' ')
-            if input().lower() == 'y':
-                remote_url = call(
-                    'git', 'config', '--get', 'remote.origin.url',
-                    capture_output=True).stdout.strip()
-                package_name = remote_url.rsplit('/')[-1].removesuffix('.git')
-                success = set_branch_protection(
-                    package_name, self.path / '.meta.toml')
-                if success:
-                    print('Successfully updated branch protection rules.')
-                else:
-                    abort(-1)
+            if self.args.admin:
+                print()
+                print('If you are an admin and are logged in via `gh auth'
+                      ' login`')
+                print('update branch protection rules? (y/N)?', end=' ')
+                if input().lower() == 'y':
+                    success = set_branch_protection(
+                        get_package_name(), self.path / '.meta.toml')
+                    if success:
+                        print('Successfully updated branch protection rules.')
+                    else:
+                        abort(-1)
             print()
             print('If everything went fine up to here:')
             if updating: