+++ CHANGES.rst
@@ -7,6 +7,8 @@ Change log
 - Add script ``bin/switch-to-pep420`` to convert a package from the old
   namespace package layout to the new PEP 420 native layout.
 
+- Add ability to run ``bin/update-python-support`` in CI.
+
 
 1.1 (2025-01-29)
 ----------------