+++ CHANGES.rst
@@ -7,6 +7,7 @@ Change log
 - Add script ``bin/switch-to-pep420`` to convert a package from the old
   namespace package layout to the new PEP 420 native layout.
 
+- Add ``pyupgrade-exclude`` to ``[pre-commit]`` section in ``.meta.toml``.
 
 1.1 (2025-01-29)
 ----------------