+++ CHANGES.rst
@@ -18,6 +18,9 @@ Change log
 
 - Add ``pyupgrade-exclude`` to ``[pre-commit]`` section in ``.meta.toml``.
 
+- Downgrade ``zc.buildout`` version pin to ``<4.1`` due to
+  ongoing setuptools compatibility issues.
+
 1.1 (2025-01-29)
 ----------------